+++ src/views/search-view.ts
@@ -15,9 +15,16 @@ interface Props {
 export function searchView(props: Props) {
   return /*html*/ `
     <h1>Search ${escape(props.query)}</h1>
-    <div>
+    <div class="results">
       ${fastMapJoin(props.results, result => /*html*/`
-        <code>${JSON.stringify(result, null, 2)}</code>
+        <div class="result-item">
+          <h6 class="result-item-name">${escape(result.name)}</h6>
+          <div class="result-item-categories">
+            ${fastMapJoin(result.categories, category => /*html*/`
+              <span class="result-item-category">${escape(category.name)}</span>
+            `)}
+          </div>
+        </div>
       `)}
     </div>
   `