+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 26 Aug 2019
- - [x] Add simple styles for header
+ - [x] Add simple styles for header & search results
  - [x] Fix `elasticsearch` config
 
 ## 24 Aug - 25 Aug, 2019